Sysanalyses
Falta validar formulários

+++ CSI477-SI-06-2018-01-atividade-002-pratica-php/sysanalyses/controller/AdminController.php
@@ -35,6 +35,31 @@ class AdminController {
 		include './view/admin/list_user.php';
 	}
 
+	public function create_user(){
+		session_start();
+		include './view/admin/create_user.php';
+	}
+
+	public function save_user($request){
+		// Recuperando dados do form
+		$name = $request['name'];
+		$email = $request['email'];
+		$password = $request['password'];
+		$type = $request['type'];
+
+		//Acesso ao Model
+		$proc = new User();
+		$result = $proc->create($name, $email, $password, $type);
+
+		if($result){
+			echo "<script>alert('Usuário criado com sucesso!');</script>";
+			header("refresh:0; url=./router_admin.php?op=2");
+		}else{
+			echo "<script>alert('Ocorreu um erro!');</script>";
+			header("refresh:0; url=./router_admin.php?op=2");
+		}
+	}
+
 	public function edit_user($lista, $id){
 		$user = new User();
 		$result = $user->editar($id);
@@ -49,17 +74,66 @@ class AdminController {
 		$result = $user->update($request);
 
 		if($result){
-			header("Location: ./router_admin.php?op=2");
+			echo "<script>alert('Usuário atualizado com sucesso!');</script>";
+			header("refresh:0; url=./router_admin.php?op=2");
 		}else{
-			echo "ERRO";
+			echo "<script>alert('Ocorreu um erro!');</script>";
+			header("refresh:0; url=./router_admin.php?op=2");
 		}
 	}
 
 	public function delete_user($id){
-		$user = new User();
-		$result = $user->delete($id);
+		//Verificar se existe exame marcados pelo usuário
+		$test = new Test();
+		$verifica = $test->verifica_delete_user($id);
 
-		header("Location: ./router_admin.php?op=2");
+		if(!$verifica){
+			$user = new User();
+			$result = $user->delete($id);
+
+			if($result){
+				echo "<script>alert('Usuário deletado com sucesso!');</script>";
+				header("Location: ./router_admin.php?op=2");
+			} else{
+				echo "<script>alert('Ocorreu um erro!');</script>";
+				header("refresh:0; url=./router_admin.php?op=2");
+			}
+
+		} else{
+			echo "<script>alert('Usuário não pode ser deletado!');</script>";
+			header("refresh:0; url=./router_admin.php?op=2");
+		}
+	}
+
+	public function create_procedure(){
+		session_start();
+		include './view/admin/create_procedure.php';
+	}
+
+	public function save_procedure($request){
+		// Recuperando dados do form
+		$name = $request['name'];
+		$price = $request['price'];
+
+		//Acesso ao Model
+
+		//Verifica se procedimento com mesmo nome existe
+		$proc = new Procedure();
+		$consulta = $proc->verifica($name);
+
+		if($consulta == 0){
+			$result = $proc->create($name, $price);
+			if($result){
+				echo "<script>alert('Procedimento salvo com sucesso!');</script>";
+				header("refresh:0; url=./router_admin.php?op=1");
+			}else{
+				echo "<script>alert('Ocorreu um erro!');</script>";
+				header("refresh:0; url=./router_admin.php?op=1");
+			}
+		} else {
+			echo "<script>alert('Procedimento já existe!');</script>";
+			header("refresh:0; url=./router_admin.php?op=1");
+		}
 	}
 
 	public function select_procedure($id){
@@ -78,21 +152,36 @@ class AdminController {
 		$result = $proc->update($request);
 
 		if($result){
-			header("Location: ./router_admin.php?op=1");
+			echo "<script>alert('Procedimento atualizado com sucesso!');</script>";
+			header("refresh:0; url=./router_admin.php?op=1");
 		}else{
-			echo "ERRO";
+			echo "<script>alert('Ocorreu um erro!');</script>";
+			header("refresh:0; url=./router_admin.php?op=1");
 		}
 	}
 
 	public function delete_procedure($id){
-		$proc = new Procedure();
-		$result = $proc->delete($id);
+		//Verificar se existe exame marcados
+		$test = new Test();
+		$result = $test->verifica_delete_procedure($id);
 
-		if($result){
-			header("Location: ./router_admin.php?op=1");
-		}else{
-			echo "ERRO";
+		if(!$result){
+			$proc = new Procedure();
+			$result = $proc->delete($id);
+			
+			if($result){
+				echo "<script>alert('Procedimento deletado com sucesso!');</script>";
+				header("Location: ./router_admin.php?op=1");
+			} else{
+				echo "<script>alert('Ocorreu um erro!');</script>";
+				header("refresh:0; url=./router_admin.php?op=1");
+			}
+
+		} else{
+			echo "<script>alert('Procedimento não pode ser deletado!');</script>";
+			header("refresh:0; url=./router_admin.php?op=1");
 		}
+		
 	}
 
 	public function tests(){
@@ -104,9 +193,5 @@ class AdminController {
   	// Invocar a View
 		include './view/admin/list_tests.php';
 		
-	}
-
-	public function relatorios(){
-		echo "CRUD RELATORIOS";
 	}
 }
